+++ iris/service/content/petition/confirmation.py
@@ -36,6 +36,17 @@ class SMSBaseHandler(Handler):
         return Petition.get(confirmation.data['petition'])
 
     @classmethod
+    def trust_user_mobile(cls, user, mobile):
+        """Manage the users mobile_trusted flag
+
+        If there is a user and the users mobil is the same as the provided
+        mobile then the mobile_trusted flag of the user is set to true.
+        """
+        if user and not user.mobile_trusted and user.mobile == mobile:
+            user.mobile_trusted = True
+            user.store(refresh=True)
+
+    @classmethod
     def handle_confirmation(cls, request, petition, token):
         confirmation = Confirmation.get_active_context_id(
             cls.build_context_id(petition, token)
@@ -87,7 +98,11 @@ class PetitionSMSHandler(SMSBaseHandler, rest.RESTMapper):
         """Confirms the mobile number on the petition
 
         If the mobile number on the owner relation matches the mobile number
-        of this conrimation the mobile_trusted flag is set to true.
+        of this confimation the mobile_trusted flag is set to true.
+
+        If the mobile on the relation is the same as the mobile of the
+        references user then the users mobile_trusted flag is also set to
+        true.
         """
         if petition is None:
             petition = self._petition(confirmation)
@@ -96,6 +111,7 @@ class PetitionSMSHandler(SMSBaseHandler, rest.RESTMapper):
             raise ValueError('Mobile number not matching')
         petition.owner = {"mobile_trusted": True}
         petition.store(refresh=True)
+        self.trust_user_mobile(petition.owner(), mobile)
 
 
 class SupportSMSHandler(SMSBaseHandler, rest.RESTMapper):
@@ -132,7 +148,21 @@ class SupportSMSHandler(SMSBaseHandler, rest.RESTMapper):
         pass
 
 
-class PetitionEMailConfirmHandler(Handler, rest.RESTMapper):
+class EMailBaseHandler(Handler):
+
+    @classmethod
+    def trust_user_email(cls, user, email):
+        """Manage the users email_trusted flag
+
+        If there is a user and the users email is the same as the provided
+        email then the email flag of the user is set to true.
+        """
+        if user and not user.email_trusted and user.email == email:
+            user.email_trusted = True
+            user.store(refresh=True)
+
+
+class PetitionEMailConfirmHandler(EMailBaseHandler, rest.RESTMapper):
     """Email confirmation handler for petitions
     """
 
@@ -204,6 +234,7 @@ class PetitionEMailConfirmHandler(Handler, rest.RESTMapper):
             raise ValueError('EMail not matching')
         petition.owner = {"email_trusted": True}
         petition.store(refresh=True)
+        self.trust_user_email(petition.owner(), email)
 
     def _petition(self, confirmation):
         return Petition.get(confirmation.data['petition'])
@@ -214,7 +245,7 @@ class PetitionEMailConfirmHandler(Handler, rest.RESTMapper):
         return fe['domain'] + fe['petition-email-confirmpath'] + '?key=' + key
 
 
-class SupportEMailConfirmHandler(Handler, rest.RESTMapper):
+class SupportEMailConfirmHandler(EMailBaseHandler, rest.RESTMapper):
     """Email confirmation handler for supporters
     """
 
@@ -262,6 +293,7 @@ class SupportEMailConfirmHandler(Handler, rest.RESTMapper):
             raise ValueError('EMail not matching')
         supporter.user = {"email_trusted": True}
         supporter.store(refresh=True)
+        self.trust_user_email(supporter.user(), email)
 
     def _petition(self, confirmation):
         return Petition.get(confirmation.data['petition'])